requirements.txt
flask
requests
python-dotenv

# ----------- Locked versions below if latest versions fail to function.
# flask==3.1.0
# requests==2.31.0
# python-dotenv==1.0.1
